@@ -16,18 +16,19 @@
 		"test": "npm run test:unit -- --run"
 	},
 	"devDependencies": {
-		"@apollo/client": "^3.13.8",
 		"@auth/sveltekit": "^1.10.0",
-		"@eslint/compat": "^1.3.1",
-		"@eslint/js": "^9.31.0",
+		"@bufbuild/protobuf": "^2.6.3",
+		"@eslint/compat": "^1.3.2",
+		"@eslint/js": "^9.33.0",
 		"@spotify/web-api-ts-sdk": "^1.2.0",
-		"@sveltejs/adapter-auto": "^6.0.1",
-		"@sveltejs/kit": "^2.25.1",
-		"@sveltejs/vite-plugin-svelte": "^6.1.0",
-		"@tailwindcss/vite": "^4.1.11",
-		"@zerodevx/svelte-toast": "^0.9.6",
+		"@sveltejs/adapter-auto": "^6.1.0",
+		"@sveltejs/kit": "^2.31.1",
+		"@sveltejs/vite-plugin-svelte": "^6.1.2",
+		"@tailwindcss/vite": "^4.1.12",
+		"@types/websocket": "^1.0.10",
+		"bits-ui": "^2.9.3",
 		"clsx": "^2.1.1",
-		"eslint": "^9.31.0",
+		"eslint": "^9.33.0",
 		"eslint-config-prettier": "^10.1.8",
 		"eslint-plugin-svelte": "^3.11.0",
 		"globals": "^16.3.0",
@@ -36,13 +37,16 @@
 		"prettier": "^3.6.2",
 		"prettier-plugin-svelte": "^3.4.0",
 		"prettier-plugin-tailwindcss": "^0.6.14",
-		"svelte": "^5.36.10",
-		"svelte-check": "^4.3.0",
+		"svelte": "^5.38.1",
+		"svelte-check": "^4.3.1",
+		"svelte-sonner": "^1.0.5",
 		"tailwind-merge": "^3.3.1",
-		"tailwindcss": "^4.1.11",
-		"typescript": "^5.8.3",
-		"typescript-eslint": "^8.37.0",
-		"vite": "^7.0.5",
+		"tailwind-variants": "^2.1.0",
+		"tailwindcss": "^4.1.12",
+		"typescript": "^5.9.2",
+		"typescript-eslint": "^8.39.1",
+		"uuid": "^11.1.0",
+		"vite": "^7.1.2",
 		"vitest": "^3.2.4",
 		"websocket": "^1.0.35"
 	},
