@@ -17,17 +17,8 @@ const tokens_default: SpotifyTokens = {
     created_at: 0,
 } as const;
 
-//                     ________  ________  ________  _________  ___  ________ ___    ___ ___  ___  ________  ________   ________  ___       _______   ________
-//                    |\   ____\|\   __  \|\   __  \|\___   ___\\  \|\  _____\\  \  /  /|\  \|\  \|\   __  \|\   ___  \|\   ___ \|\  \     |\  ___ \ |\   __  \
-//                    \ \  \___|\ \  \|\  \ \  \|\  \|___ \  \_\ \  \ \  \__/\ \  \/  / | \  \\\  \ \  \|\  \ \  \\ \  \ \  \_|\ \ \  \    \ \   __/|\ \  \|\  \
-//                     \ \_____  \ \   ____\ \  \\\  \   \ \  \ \ \  \ \   __\\ \    / / \ \   __  \ \   __  \ \  \\ \  \ \  \ \\ \ \  \    \ \  \_|/_\ \   _  _\
-//                      \|____|\  \ \  \___|\ \  \\\  \   \ \  \ \ \  \ \  \_| \/  /  /   \ \  \ \  \ \  \ \  \ \  \\ \  \ \  \_\\ \ \  \____\ \  \_|\ \ \  \\  \|
-//                        ____\_\  \ \__\    \ \_______\   \ \__\ \ \__\ \__\__/  / /      \ \__\ \__\ \__\ \__\ \__\\ \__\ \_______\ \_______\ \_______\ \__\\ _\
-//                       |\_________\|__|     \|_______|    \|__|  \|__|\|__|\___/ /        \|__|\|__|\|__|\|__|\|__| \|__|\|_______|\|_______|\|_______|\|__|\|__|
-//                       \|_________|                                       \|___|/
-
 export class SpotifyHandler {
-    public static readonly BACK_API = `${env.PUBLIC_SERVER_ADDR_DEV}/sharify`; // TODO: handle public server addr
+    public static readonly BACK_API = `${env.PUBLIC_SERVER_ADDR_DEV}/v1`; // TODO: handle public server addr
 
     private sdk: SpotifyApi | null = null;
     private tokens = tokens_default;
