@@ -2,7 +2,8 @@ import websocket from "websocket";
 import { type Writable, writable } from "svelte/store";
 import { env } from "$env/dynamic/public";
 
-import type { PartyClientID, PartyID } from "./types";
+import type { Room } from "$/lib/proto/room";
+import { bytes_to_uuid_str } from "./utils";
 
 const ws: Writable<websocket.w3cwebsocket | null> = writable(null);
 
@@ -23,15 +24,15 @@ function default_onmessage(message: websocket.IMessageEvent) {
 }
 
 export function init_ws(
-    party_id: PartyID,
-    client_id: PartyClientID,
+    room_id: Room["id"],
+    user_id: Room["users"][number]["id"],
     onopen: typeof default_onopen = default_onopen,
     onclose: typeof default_onclose = default_onclose,
     onerror: typeof default_onerror = default_onerror,
     onmessage: typeof default_onmessage = default_onmessage,
 ) {
     const _ws = new websocket.w3cwebsocket(
-        `${env.PUBLIC_SERVER_ADDR_DEV}/sharify_ws/${party_id}/${client_id}`, // TODO: Change me on prod
+        `${env.PUBLIC_SERVER_ADDR_DEV}/v1/${bytes_to_uuid_str(room_id)}/${user_id}`, // TODO prod: Change me
         undefined,
         undefined,
         undefined,
