@@ -1,7 +1,7 @@
 <script lang="ts">
     import { onMount, hasContext, getContext } from "svelte";
     import { goto } from "$app/navigation";
-    import { toast } from "@zerodevx/svelte-toast";
+    import { toast } from "svelte-sonner";
     import { Button } from "$/components/ui/button";
     import { LoaderCircle } from "lucide-svelte";
     import type { Writable } from "svelte/store";
@@ -13,7 +13,7 @@
     import CustomButton from "$/components/button.svelte";
     import { JOIN_PARTY } from "$/lib/queries";
     import type { Party } from "$/lib/types";
-    import { get_storage_value, set_storage_value } from "$/lib/utils";
+    import { bytes_to_uuid_str, get_storage_value, set_storage_value } from "$/lib/utils";
 
     if (!hasContext("GQL_Client")) {
         throw new Error("Unexpected error: Unable to get GraphQL client on context, please contact Snoupix");
@@ -33,7 +33,7 @@
 
     onMount(async () => {
         if (!data || !data.party) {
-            toast.push("Error: Party not found");
+            toast("Error: Party not found");
             return await goto("/");
         }
 
@@ -54,7 +54,7 @@
         const user_id = get_storage_value("user_id");
 
         if (user_id == null) {
-            toast.push("Unexpected error: You're not logged in");
+            toast("Unexpected error: You're not logged in");
             return await goto("/");
         }
 
@@ -62,17 +62,17 @@
             mutation: JOIN_PARTY,
             variables: { id: data.party.id, username, user_id },
         });
-        const party: (Party & { __typename: "Party" }) | { __typename: "PartyError"; error: string } | undefined =
+        const room: (Party & { __typename: "Party" }) | { __typename: "PartyError"; error: string } | undefined =
             result?.data?.joinParty;
 
-        if (!party || party.__typename == "PartyError") {
-            error = `There was an error joining the party (${party?.error ?? "Unknown error"})`;
+        if (!room || room.__typename == "PartyError") {
+            error = `There was an error joining the party (${room?.error ?? "Unknown error"})`;
             console.error(result);
             is_loading = false;
             return;
         }
 
-        const party_client = party.clients.find(c => c.username == username);
+        const party_client = room.clients.find(c => c.username == username);
 
         if (party_client == undefined) {
             error = "Username not found on party clients, check the console for more details"; // TODO: Do better
@@ -81,9 +81,9 @@
             return;
         }
 
-        set_storage_value({ user: party_client, current_room: party });
-        toast.push(`You successfully joined room "${party.name}"!`);
-        await goto(`/room/${party.id}`);
+        set_storage_value({ user: party_client, current_room: room });
+        toast(`You successfully joined room "${room.name}"!`);
+        await goto(`/room/${bytes_to_uuid_str(room.id)}`);
     }
 </script>
 
